bot/requirements.txt
python-dotenv
aiogram